+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v5.0.3 - 2023-08-03
+
+- The Modbus TCP server now returns the received unit identifier even when its own unit identifier is set to zero (the default) (solves #93).
+- The protected methods `AddUnit()` and `RemoveUnit()` have been made public.
+
 ## v5.0.2 - 2022-11-14
 
 ### Bugs Fixed